# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ChapterProgress": {
        "properties": {
          "curriculum": {
            "$ref": "#/components/schemas/CurriculumProgress"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "memory": {
            "$ref": "#/components/schemas/MemoryProgress"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "topics": {
            "items": {
              "$ref": "#/components/schemas/TopicProgress"
            },
            "title": "Topics",
            "type": "array"
          }
        },
        "required": [
          "id",
          "title",
          "curriculum",
          "memory",
          "topics"
        ],
        "title": "ChapterProgress",
        "type": "object"
      },
...
      "ConceptProgress": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "memory": {
            "$ref": "#/components/schemas/MemoryProgress"
          },
          "misconceptions": {
            "items": {
              "type": "string"
            },
            "title": "Misconceptions",
            "type": "array"
          },
          "study_state": {
            "$ref": "#/components/schemas/StudyState"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "study_state",
          "memory",
          "misconceptions"
        ],
        "title": "ConceptProgress",
        "type": "object"
      },
...
      "CourseProgress": {
        "properties": {
          "chapters": {
            "items": {
              "$ref": "#/components/schemas/ChapterProgress"
            },
            "title": "Chapters",
            "type": "array"
          },
          "course_id": {
            "format": "uuid",
            "title": "Course Id",
            "type": "string"
          },
          "curriculum": {
            "$ref": "#/components/schemas/CurriculumProgress"
          },
          "memory": {
            "$ref": "#/components/schemas/MemoryProgress"
          },
          "review_load": {
            "$ref": "#/components/schemas/ReviewLoad"
          }
        },
        "required": [
          "course_id",
          "curriculum",
          "memory",
          "review_load",
          "chapters"
        ],
        "title": "CourseProgress",
        "type": "object"
      },
...
      "CourseSummary": {
        "properties": {
          "active_concepts": {
            "title": "Active Concepts",
            "type": "integer"
          },
          "course_id": {
            "format": "uuid",
            "title": "Course Id",
            "type": "string"
          },
          "mastery": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mastery"
          },
          "review_load": {
            "$ref": "#/components/schemas/ReviewLoad"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "weak_concepts": {
            "items": {
              "$ref": "#/components/schemas/WeakConcept"
            },
            "title": "Weak Concepts",
            "type": "array"
          }
        },
        "required": [
          "course_id",
          "title",
          "review_load",
          "active_concepts",
          "mastery",
          "weak_concepts"
        ],
        "title": "CourseSummary",
        "type": "object"
      },
...
      "CurriculumProgress": {
        "description": "How far through the material the user is (study states, spec \u00a722). Says nothing about\nmemory.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "integer"
          },
          "completed": {
            "title": "Completed",
            "type": "integer"
          },
          "concepts": {
            "title": "Concepts",
            "type": "integer"
          },
          "not_studied": {
            "title": "Not Studied",
            "type": "integer"
          },
          "paused": {
            "title": "Paused",
            "type": "integer"
          },
          "studied": {
            "title": "Studied",
            "type": "integer"
          }
        },
        "required": [
          "concepts",
          "not_studied",
          "studied",
          "active",
          "paused",
          "completed"
        ],
        "title": "CurriculumProgress",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HomeSummary": {
        "description": "The signed-in user's day, per Course. `totals` sums the Courses' review loads; nothing\nelse is combined across Courses.",
        "properties": {
          "courses": {
            "items": {
              "$ref": "#/components/schemas/CourseSummary"
            },
            "title": "Courses",
            "type": "array"
          },
          "totals": {
            "$ref": "#/components/schemas/ReviewLoad"
          }
        },
        "required": [
          "totals",
          "courses"
        ],
        "title": "HomeSummary",
        "type": "object"
      },
...
      "MemoryProgress": {
        "description": "How well the trained material is retained (spec \u00a723, \u00a752). `mastery` is an estimate\n(0-1), not a measurement: the mean over trained items of level / top level, discounted\nfor lapses. Null when nothing is trained yet.",
        "properties": {
          "items_trained": {
            "title": "Items Trained",
            "type": "integer"
          },
          "learning": {
            "title": "Learning",
            "type": "integer"
          },
          "marked_hard": {
            "title": "Marked Hard",
            "type": "integer"
          },
          "mastered": {
            "title": "Mastered",
            "type": "integer"
          },
          "mastery": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mastery"
          },
          "new": {
            "title": "New",
            "type": "integer"
          },
          "relearning": {
            "title": "Relearning",
            "type": "integer"
          },
          "review": {
            "title": "Review",
            "type": "integer"
          }
        },
        "required": [
          "items_trained",
          "new",
          "learning",
          "review",
          "relearning",
          "mastered",
          "marked_hard",
          "mastery"
        ],
        "title": "MemoryProgress",
        "type": "object"
      },
...
      "ReviewLoad": {
        "description": "Upcoming reviews of items that can be reviewed now (active, trained, not paused), by the\nuser's calendar day. Buckets don't overlap: everything due today is due_now + later_today.",
        "properties": {
          "due_now": {
            "title": "Due Now",
            "type": "integer"
          },
          "later": {
            "title": "Later",
            "type": "integer"
          },
          "later_today": {
            "title": "Later Today",
            "type": "integer"
          },
          "new_to_learn": {
            "title": "New To Learn",
            "type": "integer"
          },
          "next_7_days": {
            "title": "Next 7 Days",
            "type": "integer"
          },
          "overdue": {
            "title": "Overdue",
            "type": "integer"
          },
          "tomorrow": {
            "title": "Tomorrow",
            "type": "integer"
          }
        },
        "required": [
          "due_now",
          "overdue",
          "later_today",
          "tomorrow",
          "next_7_days",
          "later",
          "new_to_learn"
        ],
        "title": "ReviewLoad",
        "type": "object"
      },
...
      "StudyState": {
        "description": "docs/PROJECT_SPEC.md \u00a722 \u2014 does NOT represent memory strength.",
        "enum": [
          "NOT_STUDIED",
          "STUDIED",
          "ACTIVE",
          "PAUSED",
          "COMPLETED"
        ],
        "title": "StudyState",
        "type": "string"
      },
...
      "TopicProgress": {
        "properties": {
          "concepts": {
            "items": {
              "$ref": "#/components/schemas/ConceptProgress"
            },
            "title": "Concepts",
            "type": "array"
          },
          "curriculum": {
            "$ref": "#/components/schemas/CurriculumProgress"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "memory": {
            "$ref": "#/components/schemas/MemoryProgress"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "curriculum",
          "memory",
          "concepts"
        ],
        "title": "TopicProgress",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WeakConcept": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "lapses": {
            "title": "Lapses",
            "type": "integer"
          },
          "marked_hard": {
            "title": "Marked Hard",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "lapses",
          "marked_hard"
        ],
        "title": "WeakConcept",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Adaptive AI Learning Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/courses/{course_id}/progress": {
      "get": {
        "operationId": "course_progress_api_v1_courses__course_id__progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "utc_offset_minutes",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 840,
              "minimum": -840,
              "title": "Utc Offset Minutes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseProgress"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Course Progress",
        "tags": [
          "progress"
        ]
      }
    },
...
    "/api/v1/courses/{course_id}/review-load": {
      "get": {
        "operationId": "review_load_api_v1_courses__course_id__review_load_get",
        "parameters": [
          {
            "in": "path",
            "name": "course_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Course Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "utc_offset_minutes",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 840,
              "minimum": -840,
              "title": "Utc Offset Minutes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewLoad"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Load",
        "tags": [
          "progress"
        ]
      }
    },
...
    "/api/v1/home": {
      "get": {
        "description": "The Home screen in one call: each of the user's Courses with its review load, active\nConcepts, mastery estimate and weakest Concepts.",
        "operationId": "home_api_v1_home_get",
        "parameters": [
          {
            "in": "query",
            "name": "utc_offset_minutes",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 840,
              "minimum": -840,
              "title": "Utc Offset Minutes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Home",
        "tags": [
          "progress"
        ]
      }
    },
...
  }
}